Move deposit offers test to separate file
