+++ vms/platformvm/service_test.go
@@ -61,13 +61,6 @@ var (
 		0xb2, 0xbc, 0x5c, 0xcf, 0x55, 0x8d, 0x80, 0x27,
 	}
 
-	genesisDepositOffer = depositOffer{
-		InterestRateNominator: uint64(0.1 * cjson.Float64(interestRateDenominator)),
-		Start:                 1659342978,
-		End:                   1672516799,
-		MinAmount:             0,
-		DepositDuration:       60,
-	}
 	// 3cb7d3842e8cee6a0ebd09f1fe884f6861e1b29c
 	// Platform address resulting from the above private key
 	testAddress = "P-testing18jma8ppw3nhx5r4ap8clazz0dps7rv5umpc36y"
@@ -839,81 +832,4 @@ func TestGetBlock(t *testing.T) {
 			assert.Equal(t, test.encoding, response.Encoding)
 		})
 	}
-}
-
-func TestGetGenesisDepositOffers(t *testing.T) {
-	service := defaultService(t)
-	service.vm.ctx.Lock.Lock()
-	defer func() {
-		err := service.vm.Shutdown()
-		assert.NoError(t, err)
-		service.vm.ctx.Lock.Unlock()
-	}()
-
-	err := genesisDepositOffer.SetID()
-	assert.NoError(t, err)
-
-	expectedReply := GetDepositOffersReply{
-		Offers: []*APIDepositOffer{
-			{
-				ID:              genesisDepositOffer.id,
-				InterestRate:    cjson.Float64(genesisDepositOffer.InterestRateFloat64()),
-				Start:           cjson.Uint64(genesisDepositOffer.Start),
-				End:             cjson.Uint64(genesisDepositOffer.End),
-				MinAmount:       cjson.Uint64(genesisDepositOffer.MinAmount),
-				DepositDuration: cjson.Uint64(genesisDepositOffer.DepositDuration),
-			},
-		},
-	}
-
-	service.vm.Clock().Set(time.Now())
-	lockRuleOffersReply := GetDepositOffersReply{}
-	err = service.GetDepositOffers(nil, nil, &lockRuleOffersReply)
-	assert.NoError(t, err)
-	assert.Equal(t, lockRuleOffersReply, expectedReply)
-}
-
-func TestGetGenesisDepositOfferById(t *testing.T) {
-	service := defaultService(t)
-	service.vm.ctx.Lock.Lock()
-	defer func() {
-		err := service.vm.Shutdown()
-		assert.NoError(t, err)
-		service.vm.ctx.Lock.Unlock()
-	}()
-
-	err := genesisDepositOffer.SetID()
-	assert.NoError(t, err)
-
-	lockRuleOfferReply := service.vm.internalState.DepositOffersChainState().GetOfferByID(genesisDepositOffer.id)
-	assert.Equal(t, *lockRuleOfferReply, genesisDepositOffer)
-}
-
-func TestAddDepositOffer(t *testing.T) {
-	service := defaultService(t)
-	service.vm.ctx.Lock.Lock()
-	defer func() {
-		err := service.vm.Shutdown()
-		assert.NoError(t, err)
-		service.vm.ctx.Lock.Unlock()
-	}()
-
-	newDepositOffer := &depositOffer{
-		InterestRateNominator: uint64(0.1 * cjson.Float64(interestRateDenominator)),
-		Start:                 1672531201,
-		End:                   1704067199,
-		MinAmount:             0,
-		DepositDuration:       120,
-	}
-
-	err := newDepositOffer.SetID()
-	assert.NoError(t, err)
-	offersState := service.vm.internalState.DepositOffersChainState()
-	newOffersState := offersState.AddOffer(newDepositOffer)
-	newOffersState.Apply(service.vm.internalState)
-	err = service.vm.internalState.Commit()
-	assert.NoError(t, err)
-
-	lockRuleOfferReply := newOffersState.GetOfferByID(newDepositOffer.id)
-	assert.Equal(t, lockRuleOfferReply, newDepositOffer)
 }
